requirements.txt
pyyaml
requests
websocket-client
